adição do backend da rota de cadastro

+++ backend/impl/src/main/java/app/tasty_flavors/usuario/repository/UsuarioRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
 
     Optional<UsuarioModel> findByEmail(String email);
